Update to build tool:  - add relation properties to doc type  - added casts array  - added ability to pass in namespace and output dir when calling console  - added Base prefix to base classes to prevent confusion on autocomplete  - added formatting to class doc notes  - added conversion tool from mysql to php for data types and added to doc notes  - added reserved words to prevent bad code from being generated

+++ src/ModelGenerator.php
@@ -118,10 +118,15 @@ class ModelGenerator
             }
 
             $result = $this->writeFile($table, $modelBase, $this->path . "/Base");
-            echo 'file written: Base' . $result['filename'] . ' - ' . $result['result'] . ' bytes' . LF;
-
-            $result = $this->writeFile($table, $model, $this->path);
             echo 'file written: ' . $result['filename'] . ' - ' . $result['result'] . ' bytes' . LF;
+
+            $filename = StringUtils::prettifyTableName($table, $this->prefix) . '.php';
+            if (!file_exists("{$this->path}/$filename")) {
+                $result = $this->writeFile($table, $model, $this->path);
+                echo 'file written: ' . $result['filename'] . ' - ' . $result['result'] . ' bytes' . LF;
+            } else {
+                echo 'file already exist: ' . $result['filename'] . ' - ' . $result['result'] . ' bytes' . LF;
+            }
         }
         echo 'done';
     }
